+++ src/Extension/TimeTrait.php
@@ -12,20 +12,20 @@ namespace DecodeLabs\Cosmos\Extension;
 use Carbon\Carbon;
 use Carbon\CarbonInterface;
 use Carbon\CarbonInterval;
-
 use DateInterval;
 use DateTime;
+use DateTimeImmutable;
+use DateTimeInterface;
 use DateTimeZone;
-
 use DecodeLabs\Cosmos;
 use DecodeLabs\Cosmos\Locale;
 use DecodeLabs\Exceptional;
-
 use IntlDateFormatter;
 use Stringable;
 
 /**
  * @template TReturn
+ * @phpstan-require-implements Time
  */
 trait TimeTrait
 {
@@ -33,7 +33,7 @@ trait TimeTrait
      * Format raw ICU date
      */
     protected function formatRawIcuDate(
-        DateTime|DateInterval|string|Stringable|int|null &$date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null &$date,
         string $pattern,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
@@ -59,7 +59,7 @@ trait TimeTrait
      * Format raw locale date
      */
     protected function formatRawLocaleDate(
-        DateTime|DateInterval|string|Stringable|int|null &$date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null &$date,
         string|int|bool|null $dateSize = true,
         string|int|bool|null $timeSize = true,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
@@ -77,7 +77,7 @@ trait TimeTrait
         }
 
         if ($hasDate && $hasTime) {
-            $wrapFormat = DateTime::W3C;
+            $wrapFormat = DateTimeInterface::W3C;
         } elseif ($hasTime) {
             $wrapFormat = 'H:i:s';
         } else {
@@ -102,7 +102,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function fullDateTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -113,7 +113,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function fullDate(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -124,7 +124,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function fullTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -136,7 +136,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function longDateTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -147,7 +147,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function longDate(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -158,7 +158,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function longTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -170,7 +170,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function mediumDateTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -181,7 +181,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function mediumDate(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -192,7 +192,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function mediumTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -204,7 +204,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function shortDateTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -215,7 +215,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function shortDate(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -226,7 +226,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function shortTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -240,7 +240,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function dateTime(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -251,7 +251,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function date(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -262,7 +262,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function time(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         string|Locale|null $locale = null
     ): mixed {
@@ -279,7 +279,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function since(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         ?bool $positive = null,
         ?int $parts = 1,
         string|Locale|null $locale = null
@@ -293,7 +293,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function sinceAbs(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         ?bool $positive = null,
         ?int $parts = 1,
         string|Locale|null $locale = null
@@ -307,7 +307,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function sinceAbbr(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         ?bool $positive = null,
         ?int $parts = 1,
         string|Locale|null $locale = null
@@ -321,7 +321,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function until(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         ?bool $positive = null,
         ?int $parts = 1,
         string|Locale|null $locale = null
@@ -335,7 +335,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function untilAbs(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         ?bool $positive = null,
         ?int $parts = 1,
         string|Locale|null $locale = null
@@ -349,7 +349,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function untilAbbr(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         ?bool $positive = null,
         ?int $parts = 1,
         string|Locale|null $locale = null
@@ -365,7 +365,7 @@ trait TimeTrait
      * @return TReturn|null
      */
     protected function formatNowInterval(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         bool $invert,
         ?int $parts,
         bool $short = false,
@@ -381,7 +381,7 @@ trait TimeTrait
      * Format interval
      */
     protected function formatRawNowInterval(
-        DateTime|DateInterval|string|Stringable|int|null &$date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null &$date,
         ?DateInterval &$interval,
         bool $invert,
         ?int $parts,
@@ -442,8 +442,8 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function between(
-        DateTime|DateInterval|string|Stringable|int|null $date1,
-        DateTime|DateInterval|string|Stringable|int|null $date2,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date1,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date2,
         ?int $parts = 1,
         string|Locale|null $locale = null
     ): mixed {
@@ -456,8 +456,8 @@ trait TimeTrait
      * @return TReturn|null
      */
     public function betweenAbbr(
-        DateTime|DateInterval|string|Stringable|int|null $date1,
-        DateTime|DateInterval|string|Stringable|int|null $date2,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date1,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date2,
         ?int $parts = 1,
         string|Locale|null $locale = null
     ): mixed {
@@ -470,8 +470,8 @@ trait TimeTrait
      * @return TReturn|null
      */
     protected function formatBetweenInterval(
-        DateTime|DateInterval|string|Stringable|int|null $date1,
-        DateTime|DateInterval|string|Stringable|int|null $date2,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date1,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date2,
         ?int $parts = 1,
         bool $short = false,
         string|Locale|null $locale = null
@@ -483,8 +483,8 @@ trait TimeTrait
      * Format interval until date
      */
     protected function formatRawBetweenInterval(
-        DateTime|DateInterval|string|Stringable|int|null &$date1,
-        DateTime|DateInterval|string|Stringable|int|null &$date2,
+        DateTimeInterface|DateInterval|string|Stringable|int|null &$date1,
+        DateTimeInterface|DateInterval|string|Stringable|int|null &$date2,
         ?DateInterval &$interval,
         ?int $parts = 1,
         bool $short = false,
@@ -525,7 +525,7 @@ trait TimeTrait
      * Prepare date for formatting
      */
     protected function prepare(
-        DateTime|DateInterval|string|Stringable|int|null $date,
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date,
         DateTimeZone|string|Stringable|bool|null $timezone = true,
         bool $includeTime = true
     ): ?DateTime {
@@ -554,10 +554,12 @@ trait TimeTrait
      * Normalize a date input
      */
     protected function normalizeDate(
-        DateTime|DateInterval|string|Stringable|int|null $date
+        DateTimeInterface|DateInterval|string|Stringable|int|null $date
     ): ?DateTime {
         if ($date === null) {
             return null;
+        } elseif ($date instanceof DateTimeImmutable) {
+            return DateTime::createFromInterface($date);
         } elseif ($date instanceof DateTime) {
             return $date;
         }